+++ module/chat.js
@@ -3,6 +3,7 @@ import { parselink } from '../lib/parselink.js'
 import { NpcInput } from '../lib/npc-input.js'
 import { Equipment } from './actor.js'
 import * as Settings from '../lib/miscellaneous-settings.js'
+import { isNiceDiceEnabled, i18n } from '../lib/utilities.js'
 
 /**
  *  This holds functions for all things chat related
@@ -56,7 +57,7 @@ export class ChatProcessor {
    * Override to return the '/help' display string
    * @param {*} isGMOnly
    */
-  help() { return "Must return help string or null" }
+  help() { return "Must return help string or null" }  // This does not need to be i18n
   
   /**
    * Override to true if this chat command only works for GMs
@@ -67,13 +68,7 @@ export class ChatProcessor {
   priv(txt) { this.registry.priv(txt) }
   pub(txt) { this.registry.pub(txt) }
   prnt(txt) { this.registry.prnt(txt) }
-  
-  // Attempt to convert original chat data into a whisper (for use when play presses SHIFT key to make roll private)
-  setWhisper() {
-    this.msgs.data.type = CONST.CHAT_MESSAGE_TYPES.WHISPER
-    this.msgs.data.whisper = [game.user.id]
-    this.msgs.event = { shiftKey: true }
-  }
+  msgs() { return this.registry.msgs }
 }
 
 class HelpChatProcessor extends ChatProcessor {
@@ -82,7 +77,7 @@ class HelpChatProcessor extends ChatProcessor {
     return line.match(/[!\/]help/i)
   }
   process(line, msgs) {
-    let t = "<a href='" + GURPS.USER_GUIDE_URL + "'>GURPS 4e Game Aid USERS GUIDE</a><br>"
+    let t = "<a href='" + GURPS.USER_GUIDE_URL + "'>" + i18n("GURPS.chatGameAidUsersGuide", 'GURPS 4e Game Aid USERS GUIDE') + "</a><br>"
     let all = ChatProcessors.processorsForAll().filter(it => !!it.help()).map(it => it.help())
     let gmonly = ChatProcessors.processorsForGMOnly().filter(it => !!it.help()).map(it => it.help())
     all.sort()
@@ -102,11 +97,7 @@ class ChatProcessorRegistry {
     this.registerProcessor(new HelpChatProcessor())
     this.msgs = { pub: [], priv: [], data: null }
   }
-  
-  _processorsForUser() {
-    return this._processors.filter(it => !it.isGMOnly() || game.user.isGM)
-  }
-  
+    
   processorsForAll() {
     return this._processors.filter(it => !it.isGMOnly())
   }
@@ -117,7 +108,7 @@ class ChatProcessorRegistry {
   // Make a pre-emptive decision if we are going to handle any of the lines in this message
   willTryToHandle(lines) {
     for (const line of lines)
-      for (const p of this._processorsForUser())
+      for (const p of this._processors)
         if (p.matches(line))
           return true
     return false
@@ -129,7 +120,8 @@ class ChatProcessorRegistry {
    */
   async processLines(lines, chatmsgData) {
     this.msgs.data = chatmsgData
-
+    delete this.msgs.event
+    
     for (const line of lines) { // use for loop to ensure single thread
       await this.processLine(line)
     } 
@@ -158,10 +150,14 @@ class ChatProcessorRegistry {
    * @returns true, if handled
    */
   async handle(line) {
-    let processor = this._processorsForUser().find(it => it.matches(line))
+    let processor = this._processors.find(it => it.matches(line))
     if (!!processor) {
-      await processor.process(line)
-      return true
+      if (processor.isGMOnly() && !game.user.isGM)
+        ui.notifications.warn(i18n("GURPS.chatYouMustBeGM"))
+      else {
+        await processor.process(line)
+        return true
+      }
     }
     return false
   }
@@ -218,6 +214,13 @@ class ChatProcessorRegistry {
     let p_setting = game.settings.get(Settings.SYSTEM_NAME, Settings.SETTING_PLAYER_CHAT_PRIVATE)
     if (game.user.isGM || p_setting) this.priv(txt)
     else this.pub(txt)
+  }
+  
+  // Attempt to convert original chat data into a whisper (for use when play presses SHIFT key to make roll private)
+  setWhisper() {
+    this.msgs.data.type = CONST.CHAT_MESSAGE_TYPES.WHISPER
+    this.msgs.data.whisper = [game.user.id]
+    this.msgs.event = { shiftKey: true }
   }
 }